Small code snippets that you might use while you are writing a big python script.
When you are writing a big scripts, sometimes you need a small functions to continue your job and finish the script. So, this resop. has many of these small snippets that you might use while writing your script. In addition, this can be helpful for beginners too, to learn such small functions and its uses.

I hope this helps!